flaskapp/requirements.txt
flask
opencv-python
tensorflow
numpy
pillow
sklearn